Split MappedArray out of fgpu.engine
Move it into a higher-level module to simplify re-use in BPipeline.
