+++ doc/katgpucbf.mapped_array.rst
@@ -1,0 +1,7 @@
+katgpucbf.mapped\_array module
+==============================
+
+.. automodule:: katgpucbf.mapped_array
+   :members:
+   :undoc-members:
+   :show-inheritance:
